Reorganized tictactoe game controllers into own package
will continue to refactor game code into discrete packages for organization
