+++ src/main/java/com/mycompany/seniorproject/games/tictactoe/TicTacToeOptionsController.java
@@ -1,0 +1,27 @@
+package com.mycompany.seniorproject.games.tictactoe;
+
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/javafx/FXMLController.java to edit this template
+ */
+
+import java.net.URL;
+import java.util.ResourceBundle;
+import javafx.fxml.Initializable;
+
+/**
+ * FXML Controller class
+ *
+ * @author Jonathan Espinal
+ */
+public class TicTacToeOptionsController implements Initializable {
+
+    /**
+     * Initializes the controller class.
+     */
+    @Override
+    public void initialize(URL url, ResourceBundle rb) {
+        // TODO
+    }    
+    
+}
