+++ src/main/resources/com/mycompany/seniorproject/TicTacToeMainMenu.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<HBox id="HBoxParent" alignment="CENTER" prefHeight="600.0" prefWidth="400.0" stylesheets="@../../../styles/tictactoemainmenu.css" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.mycompany.seniorproject.TicTacToeMainMenuController">
+<HBox id="HBoxParent" alignment="CENTER" prefHeight="600.0" prefWidth="400.0" stylesheets="@../../../styles/tictactoemainmenu.css" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.mycompany.seniorproject.games.tictactoe.TicTacToeMainMenuController">
    <children>
       <VBox alignment="CENTER" prefHeight="400.0" prefWidth="188.0">
          <children>
@@ -17,7 +17,7 @@
                   <Insets />
                </opaqueInsets>
                <padding>
-                  <Insets bottom="25.0" top="50.0" />
+                  <Insets bottom="25.0" />
                </padding>
                <font>
                   <Font size="36.0" />
